requirements.txt
Django>=6.0
psycopg2-binary>=2.9.10
djangorestframework>=3.14.0
django-filter>=23.5
Pillow>=10.1.0
requests>=2.31.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
